Grid iteration 2: <grid-col> and <grid-row>
@ungoldman had the great suggestion to just create our own css elements for
the grid (thank you!). HTML5 provides a mechanism to generate your own
custom elements (the only requirement being they have hyphens in them).
Not all browsers yet support DOM access, but they will all allow the CSS
to define and style the elements. They may not be DOM accessible in this
form, but there are polyfill JS libraries available for browsers that do
not support it yet. I may add one in as a dependency, still considering
it.

The remaining question is to see how well search engines are able to
parse this. As it will be valid HTML5, my expectation is that there
shouldn't be any major issues, but we'll see.